requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
sentence-transformers==2.3.1
pinecone-client==3.0.2
groq==0.4.1
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.4.0
plotly==5.18.0
